+++ CHANGELOG.md
@@ -30,6 +30,15 @@ The format is based on Keep a [Changelog](http://keepachangelog.com/).
 ### Breaking Changes
  - The property `Credentials` was made required in the model `CreateDbSystemDetails` in the PostgreSQL service
 
+## 107.0.0 - 2025-03-18
+### Added
+ - Support for Managed Services for Mac service
+ - Support for scheduling customer-initiated backups in the GoldenGate service
+ - Support for GPU memory clusters and fabrics in the Compute service   
+
+### Breaking Changes
+ - The properties `IpAnycastId` and `MonitorIp` were removed from the models `CreateByoipRangeDetails` and `UpdateByoipRangeDetails` in the Compute service
+
 ## 106.0.0 - 2025-03-11
 ### Added
  - Support for pipelines in the OpenSearch service